Initial commit for HU 3910 C25 Team. Working spotify login, playlist selection, and song playback.
Reworked authentication process because spotify access tokens were being saved in local storage and spotify access tokens only last for an hour.

Search bar, previous button, and next button currently do nothing (they are empty HTML elements), so it would be nice to see that changed in the near future (we'll keep it on our radar).